+++ core-ui/src/components/Clusters/views/ClusterOverview/ClusterOverviewData/ResourceCommitment/ResourceCommitment.scss
@@ -1,0 +1,30 @@
+.main {
+  width: 100%;
+  height: 40px;
+}
+
+.main div {
+  height: 100%;
+  display: inline-block;
+}
+
+.box--requests {
+  background: pink;
+}
+
+.box--limits {
+  background: magenta;
+}
+
+.box--allocatable {
+  background: red;
+}
+
+.box--allocated {
+  background: sienna;
+}
+
+.in {
+  width: 100%;
+  height: 100%;
+}
